sui_keys

Module external

Modules

__mock_MockCommandRunner

__mock_MockCommandRunner_CommandRunner

Structs

CreateKeyParams

CreateKeyResponse

External

External keystore for managing keys and aliases with an external signer.

ExternalKey

KeysResponse

MockCommandRunner

SignRequest

Request structure for signing a message with an external signer.

SignResponse

StoredKey

Enums

ExternalExecError

ProvisionMode

Provision mode requested when creating a key on an external signer.

Traits

CommandRunner