sui_keys::keystore
Trait AccountKeystore

pub trait AccountKeystore: Send + Sync {

    // Required methods
    fn import<'life0, 'async_trait>(
        &'life0 mut self,
        alias: Option<String>,
        keypair: SuiKeyPair,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn remove<'life0, 'async_trait>(
        &'life0 mut self,
        address: SuiAddress,
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn entries(&self) -> Vec<PublicKey>;
    fn export(&self, address: &SuiAddress) -> Result<&SuiKeyPair, Error>;
    fn sign_hashed<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        address: &'life1 SuiAddress,
        msg: &'life2 [u8],
    ) -> Pin<Box<dyn Future<Output = Result<Signature, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn sign_secure<'life0, 'life1, 'life2, 'async_trait, T>(
        &'life0 self,
        address: &'life1 SuiAddress,
        msg: &'life2 T,
        intent: Intent,
    ) -> Pin<Box<dyn Future<Output = Result<Signature, Error>> + Send + 'async_trait>>
       where T: Serialize + Sync + 'async_trait,
             Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn addresses_with_alias(&self) -> Vec<(&SuiAddress, &Alias)>;
    fn aliases(&self) -> Vec<&Alias>;
    fn aliases_mut(&mut self) -> Vec<&mut Alias>;
    fn get_alias(&self, address: &SuiAddress) -> Result<String, Error>;
    fn create_alias(&self, alias: Option<String>) -> Result<String, Error>;
    fn update_alias<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        old_alias: &'life1 str,
        new_alias: Option<&'life2 str>,
    ) -> Pin<Box<dyn Future<Output = Result<String, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided methods
    fn generate<'life0, 'async_trait>(
        &'life0 mut self,
        alias: Option<String>,
        generate_options: GenerateOptions,
    ) -> Pin<Box<dyn Future<Output = Result<GeneratedKey, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn addresses(&self) -> Vec<SuiAddress> { ... }
    fn alias_exists(&self, alias: &str) -> bool { ... }
    fn get_by_identity(
        &self,
        key_identity: &KeyIdentity,
    ) -> Result<SuiAddress, Error> { ... }
    fn update_alias_value(
        &mut self,
        old_alias: &str,
        new_alias: Option<&str>,
    ) -> Result<String, Error> { ... }
    fn import_from_mnemonic<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        phrase: &'life1 str,
        key_scheme: SignatureScheme,
        derivation_path: Option<DerivationPath>,
        alias: Option<String>,
    ) -> Pin<Box<dyn Future<Output = Result<SuiAddress, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }

}

Required Methods

fn import<'life0, 'async_trait>( &'life0 mut self, alias: Option<String>, keypair: SuiKeyPair, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Import a keypair into the keystore from a SuiKeyPair and optional alias.

fn remove<'life0, 'async_trait>( &'life0 mut self, address: SuiAddress, ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Remove a keypair from the keystore by its address.

fn entries(&self) -> Vec<PublicKey>

Return an array of PublicKey, consisting of every public key in the keystore.

fn export(&self, address: &SuiAddress) -> Result<&SuiKeyPair, Error>

Return SuiKeyPair for the given address.

fn sign_hashed<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, address: &'life1 SuiAddress, msg: &'life2 [u8], ) -> Pin<Box<dyn Future<Output = Result<Signature, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Sign a hash with the keypair corresponding to the given address.

fn sign_secure<'life0, 'life1, 'life2, 'async_trait, T>( &'life0 self, address: &'life1 SuiAddress, msg: &'life2 T, intent: Intent, ) -> Pin<Box<dyn Future<Output = Result<Signature, Error>> + Send + 'async_trait>>
where T: Serialize + Sync + 'async_trait, Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Sign a message with the keypair corresponding to the given address with the given intent.

fn addresses_with_alias(&self) -> Vec<(&SuiAddress, &Alias)>

Return an array of (&SuiAddress, &Alias), consisting of every address and its corresponding alias in the keystore.

fn aliases(&self) -> Vec<&Alias>

Return an array of Alias, consisting of every alias stored. Each alias corresponds to a key.

fn aliases_mut(&mut self) -> Vec<&mut Alias>

Mutable references to each alias in the keystore.

fn get_alias(&self, address: &SuiAddress) -> Result<String, Error>

Get the alias of an address.

fn create_alias(&self, alias: Option<String>) -> Result<String, Error>

Returns an alias string. Optional string can be passed, checks for duplicates

fn update_alias<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, old_alias: &'life1 str, new_alias: Option<&'life2 str>, ) -> Pin<Box<dyn Future<Output = Result<String, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Updates the alias of an existing keypair.

Provided Methods

fn generate<'life0, 'async_trait>( &'life0 mut self, alias: Option<String>, generate_options: GenerateOptions, ) -> Pin<Box<dyn Future<Output = Result<GeneratedKey, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Generate a new keypair and add it into the keystore.

fn addresses(&self) -> Vec<SuiAddress>

Return an array of SuiAddress, consisting of every address in the keystore.

fn alias_exists(&self, alias: &str) -> bool

Check if an alias exists by its name

fn get_by_identity( &self, key_identity: &KeyIdentity, ) -> Result<SuiAddress, Error>

Get address by its identity: a type which is either an address or an alias.

fn update_alias_value( &mut self, old_alias: &str, new_alias: Option<&str>, ) -> Result<String, Error>

fn import_from_mnemonic<'life0, 'life1, 'async_trait>( &'life0 mut self, phrase: &'life1 str, key_scheme: SignatureScheme, derivation_path: Option<DerivationPath>, alias: Option<String>, ) -> Pin<Box<dyn Future<Output = Result<SuiAddress, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Import from a mnemonic phrase.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.